DOMA-2027 DOMA-2030 move isObsConfigured.js to domain `common`
