+++ apps/condo/domains/common/utils/testSchema/isObsConfigured.js
