+++ apps/condo/domains/meter/schema/ExportMeterReadingsService.test.js
@@ -1,7 +1,7 @@
 /**
  * Generated by `createservice meter.ExportMeterReadingsService --type queries`
  */
-const isObsConfigured = require('@condo/domains/ticket/utils/testSchema/isObsConfigured')
+const isObsConfigured = require('@condo/domains/common/utils/testSchema/isObsConfigured')
 const { makeLoggedInAdminClient } = require('@core/keystone/test.utils')
 const { makeClientWithNewRegisteredAndLoggedInUser } = require('@condo/domains/user/utils/testSchema')
 const { makeClient } = require('@core/keystone/test.utils')
