+++ apps/condo/domains/ticket/schema/TicketAnalyticsReportService.test.js
@@ -8,7 +8,7 @@ const { makeClientWithProperty } = require('@condo/domains/property/utils/testSc
 const { createTestUser, makeLoggedInClient } = require('@condo/domains/user/utils/testSchema')
 const { createTestTicket, makeClientWithTicket, createTestTicketCategoryClassifier } = require('@condo/domains/ticket/utils/testSchema')
 const { makeClient, makeLoggedInAdminClient } = require('@core/keystone/test.utils')
-const isObsConfigured = require('@condo/domains/ticket/utils/testSchema/isObsConfigured')
+const isObsConfigured = require('@condo/domains/common/utils/testSchema/isObsConfigured')
 
 const NULL_REPLACES = {
     categoryClassifier: 'categoryClassifier not set',
